refactor: move FileAnalyzer and FileFlattener to __init__
